ResearcherScore controller and view implemented
Using the ResearcherScore view, users can now ask the system the score
of a Researcher X in the Qualis Y. The method CalculateScore() in the
controller is responsible for this. If the Researcher X has an article
that isn't avaliated in Qualis Y, you will see this information in the
view. That is my feature, half done for now. I must implement and
integrate the tests in the following commits.